Refactor stdout RedirStdout, HtmlStdout
